# public/swagger-specs/services/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ConsenSys Codefi Orchestrate API. Enables dynamic management of transactions, identities, chains, faucets and contracts.\nTransaction Requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent single transaction flows executed on the Blockchain network\nChains represent list of endpoints pointing to a Blockchain network\nFaucets represent funded accounts (holding ETH) linked to specific chains, allowed to fund newly created accounts automatically for them to be able to send transactions.\nAccounts represent Ethereum accounts (private keys). By usage of the generated cryptographic key pair, accounts can be used to sign/verify and to encrypt/decrypt messages.\nContracts represent Solidity contracts management.\nEvent Streams represent Event streams management.",
        "title": "Orchestrate API",
        "contact": {
            "name": "Contact ConsenSys Codefi Orchestrate",
            "url": "https://consensys.net/codefi/orchestrate/contact",
            "email": "orchestrate@consensys.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/jobs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of filtered jobs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Search jobs by provided filters",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of transaction hashes",
                        "name": "tx_hashes",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Chain UUID",
                        "name": "chain_uuid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Jobs found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.JobResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter in the request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates a new job as part of an already created schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Creates a new Job",
                "parameters": [
                    {
                        "description": "Job creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created Job",
                        "schema": {
                            "$ref": "#/definitions/types.JobResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single job by uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Fetch a job by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the job",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job found",
                        "schema": {
                            "$ref": "#/definitions/types.JobResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update a specific job by UUID\nWARNING: Reserved for advanced users. Orchestrate does not recommend using this endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Update job by UUID",
                "parameters": [
                    {
                        "description": "Job update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job found",
                        "schema": {
                            "$ref": "#/definitions/types.JobResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Job in invalid state for the given status update",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{uuid}/resend": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Resend transaction of specific job by UUID, effectively executing the re-sending of transaction asynchronously",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Resend Job transaction by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the job",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": ""
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{uuid}/start": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Starts a specific job by UUID, effectively executing the transaction asynchronously",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Start a Job by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the job",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": ""
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "entities.Log": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "message": {
                    "type": "string",
                    "example": "Log message"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "types.AccessTupleResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "storageKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "types.Annotations": {
            "type": "object",
            "properties": {
                "gasPricePolicy": {
                    "$ref": "#/definitions/types.GasPriceParams"
                },
                "hasBeenRetried": {
                    "type": "boolean",
                    "example": false
                },
                "oneTimeKey": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "types.CreateJobRequest": {
            "type": "object",
            "required": [
                "chainUUID",
                "scheduleUUID",
                "transaction",
                "type"
            ],
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/types.Annotations"
                },
                "chainUUID": {
                    "description": "UUID of the chain on which to create the job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "labels": {
                    "description": "List of custom labels.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "nextJobUUID": {
                    "description": "UUID of the next job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "parentJobUUID": {
                    "description": "UUID of the parent job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "scheduleUUID": {
                    "description": "UUID of the schedule on which to create the job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "transaction": {
                    "$ref": "#/definitions/types.ETHTransactionRequest"
                },
                "type": {
                    "description": "Type of job.",
                    "type": "string",
                    "example": "eth://ethereum/transaction"
                }
            }
        },
...
        "types.ETHTransactionResponse": {
            "type": "object",
            "properties": {
                "accessList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.AccessTupleResponse"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "enclaveKey": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "gas": {
                    "type": "integer"
                },
                "gasPrice": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "mandatoryFor": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "maxFeePerGas": {
                    "type": "string"
                },
                "maxPriorityFeePerGas": {
                    "type": "string"
                },
                "nonce": {
                    "type": "integer"
                },
                "privacyFlag": {
                    "type": "integer"
                },
                "privacyGroupId": {
                    "type": "string"
                },
                "privateFor": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "privateFrom": {
                    "type": "string"
                },
                "raw": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "transactionType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "types.GasPriceParams": {
            "type": "object",
            "properties": {
                "priority": {
                    "description": "Intended transaction velocity, default is `medium`. Transaction gas is adjusted depending on the velocity to optimize the mining time on user request.",
                    "type": "string",
                    "example": "very-high"
                },
                "retryPolicy": {
                    "$ref": "#/definitions/types.RetryParams"
                }
            }
        },
...
        "types.JobResponse": {
            "type": "object",
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/types.Annotations"
                },
                "chainUUID": {
                    "description": "UUID of the chain on which the job was created.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "createdAt": {
                    "description": "Date and time at which the job was created.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "labels": {
                    "description": "List of custom labels.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "logs": {
                    "description": "List of logs.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Log"
                    }
                },
                "nextJobUUID": {
                    "description": "UUID of the next job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "ownerID": {
                    "description": "ID of the job owner.",
                    "type": "string",
                    "example": "foo"
                },
                "parentJobUUID": {
                    "description": "UUID of the parent job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "scheduleUUID": {
                    "description": "UUID of the schedule on which the job was created.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                },
                "status": {
                    "description": "Status of the job.",
                    "type": "string",
                    "example": "MINED"
                },
                "tenantID": {
                    "description": "ID of the tenant executing the API.",
                    "type": "string",
                    "example": "foo"
                },
                "transaction": {
                    "$ref": "#/definitions/types.ETHTransactionResponse"
                },
                "type": {
                    "description": "Type of job.",
                    "type": "string",
                    "example": "eth://ethereum/transaction"
                },
                "updatedAt": {
                    "description": "Date and time at which the job details were updated.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "uuid": {
                    "description": "UUID of the job.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "types.RetryParams": {
            "type": "object",
            "properties": {
                "increment": {
                    "description": "Gas increment for each retry. For example, set `0.05` for a 5% increment.",
                    "type": "number",
                    "example": 0.05
                },
                "interval": {
                    "description": "Interval between different transaction retries.",
                    "type": "string",
                    "example": "2m"
                },
                "limit": {
                    "description": "Maximum gas increment over retries. For example, set `0.5` for a 50% gas increment from source transaction.",
                    "type": "number",
                    "example": 0.5
                }
            }
        },
...
        "types.UpdateJobRequest": {
            "type": "object",
            "properties": {
                "annotations": {
                    "$ref": "#/definitions/types.Annotations"
                },
                "labels": {
                    "description": "List of custom labels.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Update message.",
                    "type": "string",
                    "example": "Update message"
                },
                "receipt": {
                    "$ref": "#/definitions/ethereum.Receipt"
                },
                "status": {
                    "description": "Status of the job.",
                    "type": "string",
                    "example": "MINED"
                },
                "transaction": {
                    "$ref": "#/definitions/types.ETHTransactionRequest"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}